Open character generation from chat command
