+++ module/applications/actor/base-sheet.mjs
@@ -71,16 +71,8 @@ export default class ActorSheetEd extends ActorSheet {
     // item card description shown on item click
     html.find( ".card__name" ).click( event => this._onCardExpand( event ) );
 
-    // Character Generation
-    html.find( ".character-generation" ).click( this._onCharacterGeneration.bind( this ) );
-
     // Legend point History (Earned)
     html.find( ".legend-point__history--earned" ).click( this._onLegendPointHistoryEarned.bind( this ) );
-  }
-
-  _onCharacterGeneration( event ) {
-    event.preventDefault();
-    this.actor.characterGeneration( );
   }
 
   /**
