+++ module/documents/actor.mjs
@@ -2,8 +2,6 @@ import EdRollOptions from "../data/other/roll-options.mjs";
 import ED4E from "../config.mjs";
 import RollPrompt from "../applications/global/roll-prompt.mjs";
 import { DocumentCreateDialog } from "../applications/global/document-creation.mjs";
-import CharacterGenerationPrompt from "../applications/actor/character-generation-prompt.mjs";
-import { mapObject } from "../utils.mjs";
 
 import LegendPointHistoryEarnedPrompt from "../applications/global/legend-point-history-earned-prompt.mjs"
 
@@ -43,46 +41,6 @@ export default class ActorEd extends Actor {
     const itemDescriptionDocument = document.getElementsByClassName( "card__description" );
     const currentItemElement = itemDescriptionDocument.nextElementSibling;
     currentItemElement.classList.toggle( "d-none" )
-  }
-
-  async characterGeneration () {
-    const generation = await CharacterGenerationPrompt.waitPrompt();
-    if ( !generation ) return;
-
-    const attributeData = mapObject(
-      await generation.getFinalAttributeValues(),
-      ( attribute, value ) => [attribute, {initialValue: value}]
-    );
-    const additionalKarma = generation.availableAttributePoints;
-
-    const newActor = await this.constructor.create( {
-      name: "Rename me! I was just created",
-      type: "character",
-      system: {
-        attributes: attributeData,
-        karma: {
-          freeAttributePoints: additionalKarma,
-        },
-      },
-    } );
-
-    const namegiverDocument = await generation.namegiverDocument;
-    const classDocument = await generation.classDocument;
-    const abilities = ( await generation.abilityDocuments ).map(
-      documentData => {
-        documentData.system.source.class = namegiverDocument.uuid;
-        return documentData;
-      }
-    );
-
-    await newActor.createEmbeddedDocuments( "Item", [
-      namegiverDocument,
-      classDocument,
-      ...abilities,
-    ] );
-
-    const actorApp = newActor.sheet.render( true, {focus: true} );
-    // actorApp.activateTab("actor-notes-tab");
   }
 
   /**
